fix(pf): link pf_bindings + ParticleFlowLib into every dsl_bindings consumer; narrow field seed cast to uint32_t
C1: move PF_LIB/PF_CPP definitions before GRAPH_INT_CPP so they can be
referenced by all downstream variables. Add $(PF_CPP) to GRAPH_INT_CPP,
EXAMPLE_CPP (propagates to GALLERY/MEADOW/MEADOWCHECK/GRASSLOD/
STRESSFOREST/LIVEPROD/VIEWER_LOGIC via filter-out), TILESETBAKE_CPP,
SHLIB_CPP, RETOPO_INT_CPP, and GPU_PIPELINE_CPP. Remove the now-redundant
explicit $(PF_CPP) from TREEBAKE_CPP (it inherits via EXAMPLE_CPP) to
prevent duplicate objects on the link line.

I1: FieldConfig::seed is uint32_t; fix the parse_field cast from
static_cast<uint64_t> to static_cast<uint32_t> in pf_bindings.cpp.

+++ MatterEngine3/src/pf_bindings.cpp
@@ -125,7 +125,7 @@ bool parse_field(JSContext* c, DslState* st, JSValueConst f,
     out->kill_radius    = static_cast<float>(get_num(c, f, "killRadius", 0.1));
     out->kill_on_consume= get_bool(c, f, "killOnConsume", true);
     out->scale          = static_cast<float>(get_num(c, f, "scale", 1.0));
-    out->seed           = static_cast<uint64_t>(get_num(c, f, "seed", 0.0));
+    out->seed           = static_cast<uint32_t>(get_num(c, f, "seed", 0.0));
     out->k              = static_cast<float>(get_num(c, f, "k", 1.0));
     JSValue fade = JS_GetPropertyStr(c, f, "fade");
     if (JS_IsObject(fade)) {
